refactoring
move FetchLinkerdConfigMap and remove unused functions
