+++ web/main.go
@@ -10,6 +10,8 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/linkerd/linkerd2/pkg/config"
+
 	"github.com/linkerd/linkerd2/pkg/admin"
 	"github.com/linkerd/linkerd2/pkg/charts/linkerd2"
 	"github.com/linkerd/linkerd2/pkg/flags"
@@ -120,7 +122,7 @@ func getUUIDAndVersion(ctx context.Context, k8sAPI *k8s.KubernetesAPI, controlle
 	var uuid string
 	var version string
 
-	cm, err := healthcheck.FetchLinkerdConfigMap(ctx, k8sAPI, controllerNamespace)
+	cm, err := config.FetchLinkerdConfigMap(ctx, k8sAPI, controllerNamespace)
 	if err != nil {
 		log.Errorf("Failed to fetch linkerd-config: %s", err)
 	} else {
